Modify parsing to use struct format
The parsing is modified to use a struct of doubles instead of appending to a
vector. This makes inserting into the final array (extracted from the struct)
much easier and allows for out of order insertion into the final array.

* Add basic modifications to parse_state draft

* Start implementing Pokemon state struct as new representation of state.

* Properly use struct representation to allow for easier readability. Fix
type_id.json to properly process Pokemon types

* Delete unnecessary text files used for debugging

* Add comments to PokemonState struct and parsing functions. Decouple data
json file loading and parsing. parseState now takes in a JSON string.

Co-authored-by: Davey Hughes <davidralphhughes@gmail.com>